moved enum to proper file
