+++ client/src/components/Account/Personal/PersonalTab.tsx
@@ -1,7 +1,8 @@
 import { useState, useEffect, useRef } from 'react';
 import { useGetUserInfoQuery, useUpdateUserInfoMutation } from '../../../redux/services/user/userService';
 import { Dropdown, Spinner } from 'flowbite-react';
-import { ToastEnum, User as UserType } from '../../../types';
+import { User as UserType } from '../../../types';
+import { ToastEnum } from '../../../enums';
 import { useAppDispatch } from '../../../redux/store';
 import { updateUserInfo } from '../../../redux/feats/auth/authSlice';
 import { useToast } from '../../../utils/functions';
@@ -51,7 +52,7 @@ const PersonalTab = () => {
     try {
       const result = await updateUser(updatedFields).unwrap();
       dispatch(updateUserInfo(result));
-      showToast(ToastEnum.SAVED);
+      showToast(ToastEnum.SAVED_INFO);
     } catch (error: any) {
       console.log(error);
     }
