+++ client/src/components/Account/Profile/ProfileTab.tsx
@@ -2,7 +2,8 @@ import { useState, useEffect, useRef } from 'react';
 import { Modal, Button, Label, Spinner } from 'flowbite-react';
 import { useChangePasswordMutation } from '../../../redux/services/user/userService';
 import { useGetUserInfoQuery, useUpdateUserInfoMutation } from '../../../redux/services/user/userService';
-import { ToastEnum, User as UserType } from '../../../types';
+import { User as UserType } from '../../../types';
+import { ToastEnum } from '../../../enums';
 import ProfilePic from './ProfilePic';
 import { useAppDispatch } from '../../../redux/store';
 import {updateUserInfo} from '../../../redux/feats/auth/authSlice';
