+++ client/src/components/FriendRequest.tsx
@@ -10,7 +10,7 @@ import { User as UserType } from '../types';
 import { Spinner } from 'flowbite-react';
 import { Link } from 'react-router-dom';
 import { useToast } from '../utils/functions';
-import { ToastEnum } from '../types';
+import { ToastEnum } from '../enums';
 const FriendRequest = () => {
   const { showToast } = useToast();
   const { data, isLoading } = useGetUserFriendRequestsQuery('User');
@@ -37,7 +37,7 @@ const FriendRequest = () => {
   const acceptFriendRequestHandler = async (id: string) => {
     try {
       await acceptFriendRequest({ friendID: id });
-      showToast(ToastEnum.ACCEPTED_REQUEST);
+      showToast(ToastEnum.ACCEPTED_FRIEND_REQUEST);
     } catch (error: any) {
       console.log(error);
     }
@@ -46,7 +46,7 @@ const FriendRequest = () => {
   const rejectFriendRequestHandler = async (id: string) => {
     try {
       await rejectFriendRequest({ friendID: id });
-      showToast(ToastEnum.REJECTED_REQUEST);
+      showToast(ToastEnum.REJECTED_FRIEND_REQUEST);
     } catch (error: any) {
       console.log(error);
     }
