+++ client/src/components/Schedule/TimeSlot.tsx
@@ -3,7 +3,8 @@ import { AiFillEdit } from 'react-icons/ai';
 import { Modal, Select } from 'flowbite-react';
 import { colors } from './TimeSlotInput';
 import { ToggleSwitch } from 'flowbite-react';
-import { DaysChecked as DaysCheckedType, ToastEnum, TimeSlot as TimeSlotType } from '../../types';
+import { DaysChecked as DaysCheckedType, TimeSlot as TimeSlotType } from '../../types';
+import { ToastEnum } from '../../enums';
 import { useDeleteTimeSlotMutation, useUpdateTimeSlotMutation } from '../../redux/services/schedule/scheduleService';
 import { useGetScheduleQuery } from '../../redux/services/auth/authService';
 import { useAppSelector } from '../../redux/store';
