+++ client/src/enums.ts
@@ -7,3 +7,20 @@ export enum TypesOfInput {
   HourInput,
   MinutesInput,
 }
+
+export enum ToastEnum {
+  ACCEPTED_FRIEND_REQUEST = 'Accepted Friend Request!',
+  REJECTED_FRIEND_REQUEST = 'Rejected Friend Request.',
+  SEND_FRIEND_REQUEST = 'Friend Request Sent!',
+  CANCEL_FRIEND_REQUEST = 'Canceled Friend Request!',
+  REMOVED_FRIEND = 'Removed Friend!',
+  SAVED_INFO = "Saved Information!",
+  UPDATE_PROFILE_PICTURE = "Updated Profile Picture!",
+  UPDATE_EMAIL = "Updated Email!",
+  UPDATE_PASSWORD = "Updated Password!",
+  DELETE_ACCOUNT = "Deleted Account!",
+  CLEAR_SCHEDULE = "Cleared Schedule!",
+  CREATED_TIMESLOT = "Created Time Slot!",
+  UPDATE_TIMESLOT = "Updated Time Slot!",
+  DELETED_TIMESLOT = "Deleted Time Slot!",
+}
